feat(TabNav): move the logic into component TabNav
